requirements.txt
numpy
pandas
scikit-learn
matplotlib
jupyter
fastapi
uvicorn
scapy
flake8
pytest
requests
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
